# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WISMO API",
    "version": "1.0.0",
    "description": "Where Is My Order — AI-powered customer service for manufacturers. All endpoints require Firebase ID token authentication unless noted otherwise."
  },
  "servers": [
    { "url": "https://app.trywismo.co", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "firebaseToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Firebase ID token passed in request body as `idToken`"
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET for cron and webhook endpoints"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
...
      "TrackingResult": {
        "type": "object",
        "properties": {
          "trackingNumber": { "type": "string" },
          "carrier": { "type": "string", "enum": ["ups", "fedex", "usps", "dhl", "unknown"] },
          "status": { "type": "string" },
          "estimatedDelivery": { "type": "string" },
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string" },
                "time": { "type": "string" },
                "location": { "type": "string" },
                "description": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/agent/run": {
      "post": {
        "summary": "Run WISMO agent pipeline",
        "description": "Extract PO from customer message, look up order data (Katana + QBO), track shipment, generate AI response. Can run in test mode.",
        "tags": ["Agent"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["idToken"],
                "properties": {
                  "idToken": { "type": "string", "description": "Firebase ID token" },
                  "mode": { "type": "string", "enum": ["test", "live"], "description": "test uses most recent order" },
                  "customerEmail": { "type": "string" },
                  "customerCompany": { "type": "string" },
                  "customerMessage": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "string" },
                    "confidence": { "type": "string" },
                    "dataSources": { "type": "array", "items": { "type": "string" } },
                    "poNumber": { "type": "string", "nullable": true },
                    "orderData": { "type": "object", "nullable": true },
                    "trackingData": { "$ref": "#/components/schemas/TrackingResult" },
                    "conversationId": { "type": "string" },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid idToken" },
          "402": { "description": "Billing limit reached" },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/gmail/webhook": {
      "post": {
        "summary": "Gmail Pub/Sub webhook",
        "description": "Receives push notifications from Google Pub/Sub when new emails arrive. Processes up to 3 messages per invocation.",
        "tags": ["Gmail"],
        "security": [{ "cronSecret": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "object",
                    "properties": {
                      "data": { "type": "string", "description": "Base64-encoded JSON with emailAddress and historyId" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Processed successfully" },
          "401": { "description": "Invalid Pub/Sub token" }
        }
      }
    },
...
    "/api/gmail/callback": {
      "get": {
        "summary": "Gmail OAuth callback",
        "description": "Handles Google OAuth redirect, stores tokens, registers Gmail watch.",
        "tags": ["Gmail"],
        "parameters": [
          { "name": "code", "in": "query", "schema": { "type": "string" } },
          { "name": "state", "in": "query", "schema": { "type": "string" }, "description": "Manufacturer UID" }
        ],
        "responses": { "302": { "description": "Redirects to onboarding step 4" } }
      }
    },
...
    "/api/cron/renew-gmail-watch": {
      "get": {
        "summary": "Renew Gmail watches for all manufacturers",
        "description": "Called by Vercel Cron every 6 days. Processes in batches of 5 with retry.",
        "tags": ["Cron"],
        "security": [{ "cronSecret": [] }],
        "responses": { "200": { "description": "Renewal results" } }
      }
    },
...
  },
...
}